remove remaining trailer<Root> handling.
PDF-DOM now takes on full handling of the DOM starting at,
and including the /Root entry in the trailer. This leaves
PDF-Tools as a general purpose serialization layer.

It's now theoretically possible to target other formats that share
the same characteristics including arrays and dictionarys with byte
and stream orientated data and the need for indexing, random access,
incremental updaten and encryption.