test/refactor is-indirect method
Have moved the $.obj-num  $.gen-num and coupled $.reader attributes
from PDF::DAO::Tie to the more fundemental PDF::DAO role.

It makes sense that we may want to interogate and/or update the indirect
status of other objects. For example stream lengths are sometimes in a
following integer indirect object.