Fix "This container does not reference a native unsigned integer"
The value one assigns to a uint array slot must be unsigned. This did not throw
an error before, because Rakudo treated int and uint mostly the same. Now that
this is fixed in Rakudo, the bug in PDF::IO::Filter::RunLength surfaced. The fix
is backwards compatible (again because old Rakudo didn't care about int/uint).